Version 4.0 - Delegate Methods, Better Demo App, Improved Code, Bug Fixes
This update makes massive improvements to the flow and performance of
code, updates Objective-C compiler and build settings, redesigns the
demo app, and adds three new methods and eight new delegate methods.
-  Removed `NSNotifications` and replaced with delegate methods. This
makes integration with current projects easier and faster; it is also
more reliable. See the *Delegates* section of this document for more.
- Improved Game Center Availability Checking and Error Reporting. Now
check for the availablity of Game Center with the
`checkGameCenterAvailability` method.
- Makes improvements to thread switching - now UI tasks are performed
on the main thread. Previously they were performed on a background
thread for GameCenter.
- In depth error reporting for Game Center errors and availability
issues.
- The demo app has undergone massive improvements, including many
interface improvements. To eliminate redundancy, the iPad part of the
Demo App has been removed - Game Center Manager still works with iPad
though.
- Code cleanup and reorganization to make it easier on the eyes.
- Upgrades ARMV6 assembler codegen from THUMB to ARM